Employee rating implemented in super admin dashboard and do some cinfigurations in others
This commit implements the employee rating management feature in the Super Admin dashboard, allowing Super Admins to view, rate, and analyze employee performance with role-based access and dynamic UI components. It introduces modals for submitting and viewing ratings, AJAX-based form handling, and a color-coded rating legend. The update also includes backend route and controller configurations to support CRUD operations and JSON endpoints for ratings, as well as frontend enhancements such as progress bars, notification system, and improved interactivity for rating selection and employee summaries. Additional adjustments were made to related files and routes to ensure seamless integration and consistent user experience across the application
